% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smri_filenames.R
\name{get_smri_filenames}
\alias{get_smri_filenames}
\title{Get Structural Image Filenames}
\usage{
get_smri_filenames(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{get_image_filenames}},
\code{modalities = c("T1", "T2", "FLAIR"} so it cannot be specified}
}
\value{
Vector fo filenames
}
\description{
Return the filenames for the Structural images
}
\examples{
get_smri_filenames()
}

